SY-4392: Make CRDT counter uint32 and emit py/cpp/pb crdt outputs
The replicated-text character counter was uint64, which serializes to a
JSON string in TypeScript (bigint) and fails to decode into a Go uint64
field on the collaborative-edit dispatch path. A per-replica, per-session
character counter never approaches 2^32, so narrow it to uint32: it now
serializes as a JSON number and round-trips cleanly across the wire.

Also generate the crdt types for Python, C++, and protobuf (adding the
@py output to spatial, which crdt imports, plus hand-written __init__.py
for the new x.crdt and x.spatial packages), and carry over the Fugue
snapshot/load core and its non-interleaving + round-trip property tests.

+++ x/ts/src/crdt/types.gen.ts
@@ -28,7 +28,7 @@ export const idZ = z.object({
    * counter is the per-replica counter, incremented for every character the replica
    * inserts.
    */
-  counter: z.uint64(),
+  counter: z.uint32(),
 });
 export interface ID extends z.infer<typeof idZ> {}
